pyproject.toml
[tool.poetry]
name = "myapp-blog"
version = "2.0.0"
description = "A simple multi user blog where users can sign in, post blog posts and as well as 'Like' and 'Comment' on other posts made on the blog."
authors = ["kaushalmeena <kaushal.meena1996@gmail.com>"]
readme = "ReadMe.md"
repository = "https://github.com/kaushalmeena1996/myapp-blog"
packages = [
    { include = "blog" }
]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
Flask-SQLAlchemy = "^2.4.1"
Flask-Compress = "^1.5.0"
python-dotenv = "^0.13.0"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
flake8-black = "^0.1.1"
flake8-docstrings = "^1.5.0"
flake8-import-order = "^0.18.1"
flake8-bandit = "^2.1.2"

[tool.poetry.scripts]
start = "wsgi:start"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
